set adapter for feeds fragment
