+++ .idea/misc.xml
@@ -31,7 +31,7 @@
         <entry key="app/src/main/res/layout/fragment_bookmarked.xml" value="0.15260416666666668" />
         <entry key="app/src/main/res/layout/fragment_create.xml" value="0.16" />
         <entry key="app/src/main/res/layout/fragment_dashboard.xml" value="0.15260416666666668" />
-        <entry key="app/src/main/res/layout/fragment_feeds.xml" value="0.1375" />
+        <entry key="app/src/main/res/layout/fragment_feeds.xml" value="0.25" />
         <entry key="app/src/main/res/layout/fragment_home.xml" value="0.22" />
         <entry key="app/src/main/res/layout/fragment_my_draft.xml" value="0.15260416666666668" />
         <entry key="app/src/main/res/layout/fragment_profile.xml" value="0.1" />
