+++ app/src/main/AndroidManifest.xml
@@ -15,6 +15,11 @@
         <activity
             android:name=".individual.IndividualMainActivity"
             android:exported="true">
+            <intent-filter>
+                <action android:name="android.intent.action.MAIN" />
+
+                <category android:name="android.intent.category.LAUNCHER" />
+            </intent-filter>
 
         </activity>
         <activity
@@ -29,11 +34,7 @@
         <activity
             android:name=".splashScreen.SplashActivity"
             android:exported="true" >
-            <intent-filter>
-                <action android:name="android.intent.action.MAIN" />
 
-                <category android:name="android.intent.category.LAUNCHER" />
-            </intent-filter>
         </activity>
         <activity
             android:name=".individual.login.IndividualLoginActivity"
