+++ app/src/main/java/dev/aman/inquire/individual/IndividualMainActivity.kt
@@ -2,23 +2,13 @@ package dev.aman.inquire.individual
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
-import android.system.Os.close
-import androidx.appcompat.app.ActionBarDrawerToggle
-import androidx.drawerlayout.widget.DrawerLayout
 import androidx.fragment.app.Fragment
-import androidx.navigation.Navigation
-import androidx.navigation.findNavController
-import androidx.navigation.ui.AppBarConfiguration
-import androidx.navigation.ui.setupActionBarWithNavController
-import androidx.navigation.ui.setupWithNavController
 import com.google.android.material.bottomnavigation.BottomNavigationView
-import com.google.android.material.navigation.NavigationView
 
 import dev.aman.inquire.R
-import dev.aman.inquire.individual.data.Repository
 import dev.aman.inquire.individual.ui.bookmarked.BookmarkedFragment
 import dev.aman.inquire.individual.ui.create.CreateFragment
-import dev.aman.inquire.individual.ui.Home.HomeFragment
+import dev.aman.inquire.individual.ui.home.HomeFragment
 
 class IndividualMainActivity : AppCompatActivity() {
     private lateinit var currentFragment : Fragment
