--- app/src/main/java/dev/aman/inquire/individual/ui/Home/feeds/FeedsFragment.kt
@@ -1,23 +1,0 @@
-package dev.aman.inquire.individual.ui.Home.feeds
-
-import android.os.Bundle
-import androidx.fragment.app.Fragment
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import dev.aman.inquire.R
-
-
-class FeedsFragment : Fragment() {
-
-
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_feeds, container, false)
-    }
-
-
-    }
