+++ app/src/main/java/dev/aman/inquire/individual/ui/home/HomeFragment.kt
@@ -1,4 +1,4 @@
-package dev.aman.inquire.individual.ui.Home
+package dev.aman.inquire.individual.ui.home
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
