+++ app/src/main/java/dev/aman/inquire/individual/ui/home/HomeFragmentAdapter.kt
@@ -1,10 +1,10 @@
-package dev.aman.inquire.individual.ui.Home
+package dev.aman.inquire.individual.ui.home
 
 import androidx.fragment.app.Fragment
 import androidx.viewpager2.adapter.FragmentStateAdapter
-import dev.aman.inquire.individual.ui.Home.articles.ArticlesFragment
-import dev.aman.inquire.individual.ui.Home.feeds.FeedsFragment
-import dev.aman.inquire.individual.ui.Home.recent.RecentFragment
+import dev.aman.inquire.individual.ui.home.articles.ArticlesFragment
+import dev.aman.inquire.individual.ui.home.feeds.FeedsFragment
+import dev.aman.inquire.individual.ui.home.recent.RecentFragment
 
 class HomeFragmentAdapter(fragment: Fragment): FragmentStateAdapter(fragment){
     override fun getItemCount() = 3
