+++ app/src/main/java/dev/aman/inquire/individual/ui/home/articles/ArticlesFragment.kt
@@ -1,4 +1,4 @@
-package dev.aman.inquire.individual.ui.Home.recent
+package dev.aman.inquire.individual.ui.home.articles
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -8,23 +8,17 @@ import android.view.ViewGroup
 import dev.aman.inquire.R
 
 
-
-
-class RecentFragment : Fragment() {
-
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-
-
-    }
+class ArticlesFragment : Fragment() {
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_recent, container, false)
+        return inflater.inflate(R.layout.fragment_articles, container, false)
     }
+
+
+
 
 }
