+++ app/src/main/java/dev/aman/inquire/individual/ui/home/feeds/FeedsFragment.kt
@@ -1,0 +1,38 @@
+package dev.aman.inquire.individual.ui.home.feeds
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.recyclerview.widget.LinearLayoutManager
+import dev.aman.inquire.databinding.FragmentFeedsBinding
+import dev.aman.inquire.individual.data.DataSource
+import dev.aman.inquire.individual.ui.home.HomeAdapter
+
+
+class FeedsFragment : Fragment() {
+    private lateinit var binding: FragmentFeedsBinding
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        // Inflate the layout for this fragment
+        binding= FragmentFeedsBinding.inflate(inflater, container, false)
+        return binding.root
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+        binding.feedsRecyclerView.layoutManager=LinearLayoutManager(context)
+        val mydataset= DataSource().getData()
+        val recyclerView=binding.feedsRecyclerView
+        recyclerView.adapter= HomeAdapter(this,mydataset)
+
+
+    }
+
+
+
+
+}
