+++ app/src/main/res/layout/fragment_feeds.xml
@@ -4,8 +4,17 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".individual.ui.Home.feeds.FeedsFragment">
+    tools:context=".individual.ui.home.feeds.FeedsFragment">
 
 
-
+    <androidx.recyclerview.widget.RecyclerView
+        android:id="@+id/feeds_recycler_view"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent"
+        tools:listitem="@layout/layout_inquire"
+        />
 </androidx.constraintlayout.widget.ConstraintLayout>
