+++ app/src/main/res/layout/fragment_home.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".individual.ui.Home.HomeFragment">
+    tools:context=".individual.ui.home.HomeFragment">
 
 
 <com.google.android.material.tabs.TabLayout
