+++ app/src/main/res/layout/fragment_recent.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".individual.ui.Home.recent.RecentFragment">
+    tools:context=".individual.ui.home.recent.RecentFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
